backend/requirements.txt
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.12.3
black==24.1.1
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
datasets==4.0.0
dill==0.3.8
distro==1.9.0
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
dotenv==0.9.9
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.19.1
fitz==0.0.1.dev2
flake8==7.0.0
frontend==0.0.3
frozenlist==1.7.0
fsspec==2025.3.0
google==3.0.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gradio==5.44.0
gradio_client==1.12.1
groovy==0.1.2
groq==0.31.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
isodate==0.7.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-groq==0.3.8
langchain-openai==0.3.32
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.25
looseversion==1.3.0
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mcp==1.13.1
mdurl==0.1.2
mlx==0.28.0
mlx-lm==0.26.4
mlx-metal==0.28.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy==1.8.0
mypy_extensions==1.1.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
oauthlib==3.3.1
openai==1.106.1
opencv-python==4.10.0.84
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
pathlib==1.0.1
pathspec==0.12.1
pdf2image==1.17.0
pillow==10.2.0
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
prov==2.1.1
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.2.3
pypdf==6.0.0
pytesseract==0.3.10
pytest==8.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruff==0.12.10
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlite-vec==0.1.6
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.3
structlog==24.1.0
sympy==1.14.0
tavily-python==0.7.11
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
tomli==2.2.1
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traits==7.0.2
transformers==4.55.4
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
